# Avoid calling mget with massive number of keys in Readdir