@davies 2021-01-24
@suzaku Based on the result from benchmark, we should have a fast path for small directorym, thanks.