@davies 2021-01-24
Could you also add a benchmark in Go (for both of small and large directory)?