@suzaku 2021-01-24
@davies Is it safe here to use multiple `HSCAN` instead of a huge `HGETALL` here?