@davies 2021-01-25
@suzaku It's OK to use HSCAN. Right now, what's behavior for 5 millions files?