@suzaku 2021-01-25
If `HSCAN` is OK, we can set up a goroutine to do HSCAN on smaller batches instead of calling `HGETALL` upfront.
I haven't tested 5 million files yet. Creating files can be quite time consuming, I didn't wait that long...