@davies 2021-01-27
Let's merge this one first, then optimize the hgetall() later.